requirements.txt
Flask
flask-cors
gevent
python-dotenv
requests
